Replace resources adapter with sectioned one.
